Move log removal into own go routine
